server/requirements.txt
twilio
sendgrid
django-polymorphic
aioredis==1.2.0
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
channels==2.1.6
channels-redis==2.3.2
constantly==15.1.0
daphne==2.2.4
Django==2.1.4
django-cors-headers==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
hiredis==0.3.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.5.6
numpy==1.15.4
pandas==0.23.4
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
